cvMLAgentBaseline/agentic_mcp/requirements.txt
# LangChain and Agent Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20

# MCP Protocol
mcp>=0.1.0

# OpenAI (for LLM)
openai>=1.0.0

# MongoDB
pymongo>=4.6.0

# Redis
redis>=5.0.0

# WebSocket
websockets>=12.0

# HTTP Requests (for Cloudflare)
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Type Hints
pydantic>=2.0.0
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0

